Re-add buffer nulls
In hindsight, nulling the buffers is a better fail-safe when the IDS cannot be found. Without nulling the buffer, GetFlString would return 0 in this case and it wouldn't write any bytes. Thus the buffer would still contain random stack bytes. It would be possible to check the return value for 0 and then return an empty string, but this is less elegant.